BUG: Restore return value in notebook reprs
Monkey patches the _ipython_display_ method onto NDFrame, so that
notebook cells have a real return value. Setting the
display.html.table_schema will monkey patch the method on,
and remove it when unset.

closes https://github.com/ipython/ipython/issues/10491
