+++ pandas/core/config_init.py
@@ -307,6 +307,48 @@ def mpl_style_cb(key):
     return val
 
 
+def table_schema_cb(key):
+    # Having _ipython_display_ defined messes with the return value
+    # from cells, so the Out[x] dicitonary breaks.
+    # Currently table schema is the only thing using it, so we'll
+    # monkey patch `_ipython_displa_` onto NDFrame when config option
+    # is set
+    # see https://github.com/pandas-dev/pandas/issues/16168
+    from pandas.core.generic import NDFrame
+
+    def _ipython_display_(self):
+        from pandas.errors import UnserializableWarning
+
+        try:
+            from IPython.display import display
+        except ImportError:
+            return None
+
+        # Series doesn't define _repr_html_ or _repr_latex_
+        latex = self._repr_latex_() if hasattr(self, '_repr_latex_') else None
+        html = self._repr_html_() if hasattr(self, '_repr_html_') else None
+        try:
+            table_schema = self._repr_table_schema_()
+        except Exception as e:
+            warnings.warn("Cannot create table schema representation. "
+                          "{}".format(e), UnserializableWarning)
+            table_schema = None
+        # We need the inital newline since we aren't going through the
+        # usual __repr__. See
+        # https://github.com/pandas-dev/pandas/pull/14904#issuecomment-277829277
+        text = "\n" + repr(self)
+
+        reprs = {"text/plain": text, "text/html": html, "text/latex": latex,
+                 "application/vnd.dataresource+json": table_schema}
+        reprs = {k: v for k, v in reprs.items() if v}
+        display(reprs, raw=True)
+
+    if cf.get_option(key):
+        NDFrame._ipython_display_ = _ipython_display_
+    elif getattr(NDFrame, '_ipython_display_', None):
+        del NDFrame._ipython_display_
+
+
 with cf.config_prefix('display'):
     cf.register_option('precision', 6, pc_precision_doc, validator=is_int)
     cf.register_option('float_format', None, float_format_doc,
@@ -374,7 +416,7 @@ with cf.config_prefix('display'):
     cf.register_option('latex.multirow', False, pc_latex_multirow,
                        validator=is_bool)
     cf.register_option('html.table_schema', False, pc_table_schema_doc,
-                       validator=is_bool)
+                       validator=is_bool, cb=table_schema_cb)
 
 
 cf.deprecate_option('display.line_width',
